Refactor/address comments
